Finished the graph class
The most substantial change was a change to the graph data structure.  The
value of each node is now a dictionary instead of a list.  This permits
easier addition and deletion of nodes.

Several features have been added to the graph class:
  - The copy function was fixed.
  - The __init__ function can now cast a dictionary to a graph
  - Nodes can be added without any edges
  - Edges can be added separately
  - Nodes can be deleted
  - A function to count the number of nodes in the graph

I additionally wrote a brute force traveling salesman function, but it is
not functional because it solves the wrong problem.  (The function simply
finds the shortest path between two nodes, not the shortest path that visits
every point.)